azure_function/requirements.txt
# Azure Functions
azure-functions==1.20.0

# HTTP client
httpx==0.27.0

# Data parsing
PyYAML==6.0.1

# Logging
python-json-logger==2.0.7